update the codes

+++ dataloader/dataset_semantickitti.py
@@ -115,7 +115,7 @@ class voxel_dataset(data.Dataset):
         
         if (intervals == 0).any(): print("Zero interval!")
 
-        grid_ind = (np.floor((np.clip(xyz, min_bound, max_bound) - min_bound) / intervals)).astype(np.int)
+        grid_ind = (np.floor((np.clip(xyz, min_bound, max_bound) - min_bound) / intervals)).astype(np.int32)
 
         # process voxel position
         dim_array = np.ones(len(self.grid_size) + 1, int)
@@ -530,7 +530,7 @@ def collate_fn_BEV_tta(data):
 
 def collate_fn_BEV_ms(data):
     data2stack = np.stack([d[0] for d in data]).astype(np.float32)
-    label2stack = np.stack([d[1] for d in data]).astype(np.int)
+    label2stack = np.stack([d[1] for d in data]).astype(np.int32)
     grid_ind_stack = [d[2] for d in data]
     point_label = [d[3] for d in data]
     xyz = [d[4] for d in data]
@@ -542,7 +542,7 @@ def collate_fn_BEV_ms(data):
 
 def collate_fn_BEV_ms_tta(data):
     data2stack = np.stack([d[0] for d in data]).astype(np.float32)
-    label2stack = np.stack([d[1] for d in data]).astype(np.int)
+    label2stack = np.stack([d[1] for d in data]).astype(np.int32)
     grid_ind_stack = [d[2] for d in data]
     point_label = [d[3] for d in data]
     xyz = [d[4] for d in data]
